refactor: Apps class broken down to Apps, AppLauncher, AppSearch, and MenuDefaults
